requirements.txt
asgiref==3.7.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-ckeditor==6.5.1
django-dotenv==1.4.2
django-extensions==3.2.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-taggit==4.0.0
idna==3.4
oauthlib==3.2.2
pilkit==2.0
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.0
urllib3==2.0.2
